% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar-press.R
\name{bar_press}
\alias{bar_press}
\title{Barometric/total pressure (p)}
\usage{
bar_press(alt = 0)
}
\arguments{
\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of barometric pressure [kPa].
}
\description{
Calculates barometric pressure based on altitude. This is also called total pressure.
Eq.3 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
# Calculate barometric pressure at sea level
bar_press()
# Calculate barometric pressure 1000 m above sea level
bar_press(alt = 1000)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{climate}
\keyword{indoor}

