% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewpoint.R
\name{dewpoint}
\alias{dewpoint}
\title{Dew-point temperature (t_d)}
\usage{
dewpoint(hum.ratio, alt = 0)
}
\arguments{
\item{hum.ratio}{Vector of humidity ratios [kg/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of dew-point temperatures [degC].
}
\description{
Calculates the dew-point temperature based on absolute humidity.
Eq.37 & 38 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
dewpoint(hum.ratio = 0.010)

hum.ratio = seq(0, 0.050, by = 0.001)
rel.hum = rel.hum(temp.air = dewpoint(hum.ratio = hum.ratio), hum.ratio = hum.ratio, round = F)
plot(hum.ratio, rel.hum - 100,
   xlab = "Humidity ratio [g/kg]",
   ylab = "Relative humidity difference [\%]",
   main = "Deviation from 100 \% relative humidity")
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{climate}
\keyword{indoor}

