% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enthalpy.R
\name{enthalpy}
\alias{enthalpy}
\title{Enthalpy (h)}
\usage{
enthalpy(temp.db, hum.ratio)
}
\arguments{
\item{temp.db}{Vector of dry bulb temperatures [degC].}

\item{hum.ratio}{Vector of humidity ratios [kg/kg].}
}
\value{
Returns a vector of enthalpies [kJ/kg].
}
\description{
Calculates the enthalpy of a given moist air state.
Eq. 32 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
enthalpy(25, 0.010)

# Calculating cooling demand
density <- 1.2 # density of air (kg/m3)
(enthalpy(35, 0.020) - enthalpy(15, 0.010)) * density
}
\author{
Christoffer Rasmussen
}

