% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enthalpy-intersect.R
\name{enthalpy_intersect}
\alias{enthalpy_intersect}
\title{Enthalpy and 100 \% RH intersection}
\usage{
enthalpy_intersect(enthalpy, alt = 0)
}
\arguments{
\item{enthalpy}{Vector of enthalpies [kJ/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns the dry-bulb temperature where at which the enthalpy and 100 \% RH line intersects [degC].
}
\description{
Calculates the intersection of RH 100 \% line and the enthalpy line.
}
\examples{
enthalpy_intersect(enthalpy = 50)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{internal}

