% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hum-ratio.R
\name{hum_ratio}
\alias{hum_ratio}
\title{Humidity ratio (W)}
\usage{
hum_ratio(rel.hum, temp.air, alt = 0)
}
\arguments{
\item{rel.hum}{Vector of relative humidities [\%].}

\item{temp.air}{Vector of air temperatures [degC].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of humidity ratios [kg/kg].
}
\description{
Calculates the humidity ratio from relative humidity and temperature.
Eq. 25 & 12 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
hum_ratio(rel.hum = 60, temp.air = 25, alt = 0)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{climate}
\keyword{indoor}

