% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hum-ratio-enthalpy.R
\name{hum_ratio_enthalpy}
\alias{hum_ratio_enthalpy}
\title{Humidity ratio for enthalpy lines}
\usage{
hum_ratio_enthalpy(enthalpy, temp.db, alt = 0)
}
\arguments{
\item{enthalpy}{Vector of enthalpies [kJ/kg].}

\item{temp.db}{Vector of dry bulb temperatures [degC].}
}
\value{
Returns a vector of humidity ratios [kg/kg].
}
\description{
Calculates the humidity ratio based on the enthalpy.
Eq. 32 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
hum_ratio_enthalpy(enthalpy = 30, temp.db = 25)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{internal}

