% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par-w-press.R
\name{par_w_press}
\alias{par_w_press}
\title{Partial water vapour pressure (p_w)}
\usage{
par_w_press(hum.ratio, alt = 0)
}
\arguments{
\item{hum.ratio}{Vector of humidity ratios [kg/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of partial water vapour pressure [kPa].
}
\description{
Calculates partial water vapour pressure.
Eq.36 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
par_w_press(hum.ratio = 0.010)

hum.ratio <- rep(c(0.005, 0.010), 2)
alt <- rep(c(0, 500), each = 2)
par_w_press(hum.ratio, alt)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{climate}
\keyword{indoor}

