% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-chart.R
\name{psychrometric_chart}
\alias{psychrometric_chart}
\title{Psychrometric chart}
\usage{
psychrometric_chart(temp.db = NULL, hum.ratio = NULL, temp.min = -15,
  temp.max = 30, humidity.max = 0.02, alt = 0, mollier = FALSE,
  alpha = 0.25, disable.warnings = TRUE)
}
\arguments{
\item{temp.db}{Vector of dry-bulb temperatures [degC]. Defaults to NULL.}

\item{hum.ratio}{Vector of humidity ratios [kg/kg]. Defaults to NULL.}

\item{temp.min}{Minimum value of temperature axis [degC]. Defaluts to -15 degC.}

\item{temp.max}{Maximum value of temperature axis [degC]. Defaluts to 30 degC.}

\item{humidity.max}{Maximum value of humidity axis [kg/kg]. Defaults to 0.020 kg/kg.}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}

\item{mollier}{Boolean operator. Plot Mollier chart (tx-chart) or tx-chart. Defaults to FALSE.}

\item{alpha}{Transparancy of the data points. Defaults to 0.25.}

\item{disable.warnings}{Some ignorable warnings appear when plotting the chart. To see these and eventually others disable.warnings should be FALSE. Defaults to TRUE.}
}
\value{
Plots the psychrometric chart. If the error
"TEXT_SHOW_BACKTRACE environmental variable.
Error in grid.Call(L_textBounds, as.graphicsAnnot(x$label), x$x, x$y,  :
polygon edge not found"
shows up. Try run psychrometric_chart() again.
}
\description{
Plot psychrometric chart with or with data.
}
\examples{
psychrometric_chart()

temp.db <- c(20, 23, 18, 20, 10, 27)
hum.ratio <- c(0.005, 0.009, 0.004, 0.011, 0.004, 0.01)
psychrometric_chart(temp.db, hum.ratio, alpha = 1)

psychrometric_chart(temp.db, hum.ratio, temp.min=5, mollier = TRUE, alpha = 1)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{climate}
\keyword{indoor}

