% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-hum-ratio.R
\name{sat_hum_ratio}
\alias{sat_hum_ratio}
\title{Humidity ratio at satuation (W_s)}
\usage{
sat_hum_ratio(temp.air, alt = 0)
}
\arguments{
\item{temp.air}{Vector of air temperatures [degC].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of water vapour content at satuatuation [kg/kg].
}
\description{
Eq. 23 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
sat_hum_ratio(temp.air = 25, alt = 0)
}
\author{
Christoffer Rasmussen
}

