% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope-rel-hum.R
\name{slope_rel_hum}
\alias{slope_rel_hum}
\title{Slope of relative humidity line}
\usage{
slope_rel_hum(temp.1, temp.2, rel.hum)
}
\arguments{
\item{temp.1}{Temperature of first point [degC].}

\item{temp.2}{Temperature of second point [degC].}

\item{rel.hum}{Relative humidity line to find slope of [\%].}
}
\value{
Returns the slope angle [deg].
}
\description{
Calculate slope of relative humidity line.
}
\examples{
slope_rel_hum(15, 20, 80)
}
\author{
Christoffer Rasmussen
}
\keyword{internal}

