% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-vol.R
\name{specific_vol}
\alias{specific_vol}
\title{Specific volume (v)}
\usage{
specific_vol(temp.db, hum.ratio, alt = 0)
}
\arguments{
\item{temp.db}{Vector of dry bulb temperatures [degC].}

\item{hum.ratio}{Vector of humidity ratios [kg/kg].}
}
\value{
Returns a vector of specific volumes [m3/kg].
}
\description{
Calculates the specific volume
Eq. 28 (alternative) - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
specific_vol(25, 0.010)
}
\author{
Christoffer Rasmussen
}
\keyword{climate}
\keyword{indoor}

