% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-climateeng.R
\name{theme_climateeng}
\alias{theme_climateeng}
\title{Custom ggplot2 theme}
\usage{
theme_climateeng()
}
\value{
Returns a ggplot2 theme.
}
\description{
Custom theme for climateeng plots.
}
\examples{
theme_climateeng()
}
\author{
Christoffer Rasmussen
}
\keyword{internal}

