% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-climateeng-psy.R
\name{theme_climateeng_psy}
\alias{theme_climateeng_psy}
\title{Custom ggplot2}
\usage{
theme_climateeng_psy(asp = NULL)
}
\arguments{
\item{asp}{Aspect ratio of plot. Defaults to NULL.}
}
\value{
Returns a ggplot2 theme.
}
\description{
Custom theme for climateeng psychrometric chart.
}
\examples{
theme_climateeng_psy()
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{internal}

