% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetbulb-intersect.R
\name{wetbulb_intersect}
\alias{wetbulb_intersect}
\title{Wet-bulb intersection with dry-bulb temperature axis}
\usage{
wetbulb_intersect(temp.wb, alt = 0)
}
\arguments{
\item{temp.wb}{Vector of wet-bulb temperatures [degC].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of dry-bulb temperatures at which the wet-bulb temperature line intersects [degC].
}
\description{
Calculates the drybulb temperature at which the wetbulb temperature line intersects in the psychrometric chart.
Eq.32 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
wetbulb_intersect(15)
wetbulb_intersect(15, alt = 1000)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{internal}

