% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel-hum.R
\name{rel_hum}
\alias{rel_hum}
\title{Relative humidity (phi)}
\usage{
rel_hum(temp.air, hum.ratio, alt = 0, round = TRUE)
}
\arguments{
\item{temp.air}{Vector of air temperatures [degC].}

\item{hum.ratio}{Vector of humidity ratios [kg/kg]}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}

\item{round}{Boolean operator. If TRUE (default) calculated relative humidities above satuation level will be rounded to 100 \%.}
}
\value{
Returns a vector of relative humidities [\%].
}
\description{
Calculates relative humidity.
Eq.24 - ASHRAE Fundamentals Handbook, Psychrometrics.
}
\examples{
rel_hum(temp.air = 25, hum.ratios = 0.005, alt = 500)
rel_hum(temp.air = seq(10, 20, 2), hum.ratios = 0.010)
}
\author{
Christoffer Rasmussen, MSc in Engineering
}
\keyword{climate}
\keyword{indoor}

