% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-parms.R
\name{theme_parms}
\alias{theme_parms}
\title{ggplot2 parameters}
\usage{
theme_parms()
}
\value{
Returns a list of ggplot2 theme parameters.
}
\description{
Parameters for climateeng themes.
}
\examples{
theme_parms()
}
\author{
Christoffer Rasmussen
}
\keyword{internal}

